requirements.txt
asgiref==3.6.0
Django==4.1
psycopg2==2.9.6
pytz==2021.1
sqlparse==0.4.1
djangorestframework==3.14.0